+++ lib/specdown/command/option_parser.rb
@@ -16,16 +16,28 @@ module Specdown
           Specdown::Config.root = root
         end
 
-        opts.on '-f', '--format plain|color', 'defaults to "color"' do |format|
-          case format
-            when 'plain' then Specdown::Config.reporter = Specdown::TerminalReporter
-            when 'color' then Specdown::Config.reporter = Specdown::ColorTerminalReporter
-          end
+        opts.on '-c', '--colorized', 'Colorize the terminal output' do
+          Specdown::Config.reporter = Specdown::ColorTerminalReporter
+        end
+
+        opts.on '-n', '--non-colorized', 'Display plain terminal output' do
+          Specdown::Config.reporter = Specdown::TerminalReporter
         end
 
         opts.on_tail '-h', '--help', 'Display this screen'  do
           puts opts
           exit
+        end
+
+        opts.on '-o', '--output=text|terminal', ["text", "terminal"], 'output to either the terminal or a text file "specdown_report.txt". Defaults to "terminal".' do |output_format|
+          case output_format
+            when "text"     then Specdown::Config.reporter = Specdown::TextReporter
+            when "terminal" then Specdown::Config.reporter = Specdown::ColorTerminalReporter
+          end
+        end
+
+        opts.on '-f', '--format=short|condensed', [:short, :condensed], 'length and style of output. Defaults to "short".' do
+          Specdown::Config.format = format
         end
 
         opts.on_tail "-v", "--version", "Show version" do