+++ lib/specdown/event_handlers/command_complete.rb
@@ -1,3 +1,4 @@
 Specdown::EventServer.register :command_complete do |results|
   Specdown.reporter.print_summary(results)
+  Specdown.reporter.print_end
 end